Fix the first argument to calcuateCrc() to mean hardware-suport
The changes by commit f7444a2aeb77edab694f9842b8cc2233f23ecbf4 had a bug where
the first argument to the calculateCrc() function was treated as an intention
to use the table-lookup method for CRC calculation, when it was intended to be
the opposite. This commit fixes this bug.

Thanks to @ivan for finding this bug.